Rename app and package names to "Launch Chat"
