+++ app/src/main/kotlin/org/vinaygopinath/launchchat/Constants.kt
@@ -1,0 +1,6 @@
+package org.vinaygopinath.launchchat
+
+object Constants {
+
+    const val GITHUB_REPO_URL = "https://github.com/vinaygopinath/launch-chat"
+}
