+++ app/src/main/kotlin/org/vinaygopinath/launchchat/helpers/ClipboardHelper.kt
@@ -1,4 +1,4 @@
-package org.vinaygopinath.openinchat.helpers
+package org.vinaygopinath.launchchat.helpers
 
 import android.content.ClipDescription
 import android.content.ClipboardManager
