+++ app/src/main/kotlin/org/vinaygopinath/launchchat/helpers/PhoneNumberHelper.kt
@@ -1,4 +1,4 @@
-package org.vinaygopinath.openinchat.helpers
+package org.vinaygopinath.launchchat.helpers
 
 import io.michaelrocks.libphonenumber.android.PhoneNumberUtil
 import javax.inject.Inject
