--- app/src/main/kotlin/org/vinaygopinath/openinchat/Constants.kt
@@ -1,6 +1,0 @@
-package org.vinaygopinath.openinchat
-
-object Constants {
-
-    const val GITHUB_REPO_URL = "https://github.com/vinaygopinath/open-in-chat"
-}
